+++ src/main/java/com/bookstore/service/CustomerServices.java
@@ -2,6 +2,7 @@ package com.bookstore.service;
 
 import com.bookstore.dao.CustomerDAO;
 import com.bookstore.dao.HashGenerator;
+import com.bookstore.dao.ReviewDAO;
 import com.bookstore.entity.Customer;
 import com.bookstore.store.Message;
 
@@ -17,8 +18,10 @@ public class CustomerServices {
     private HttpServletRequest request;
     private HttpServletResponse response;
     private CustomerDAO customerDAO;
+    private ReviewDAO reviewDAO;
     public CustomerServices(HttpServletRequest request, HttpServletResponse response) {
         customerDAO = new CustomerDAO();
+        reviewDAO = new ReviewDAO();
         this.request = request;
         this.response = response;
     }
@@ -153,9 +156,20 @@ public class CustomerServices {
             Message message = new Message("Could not delete customer", "Could not find customer", "error");
             request.setAttribute("message", message);
         }else{
-            customerDAO.delete(customerId);
-            Message message = new Message("Delete successful", "Delete customer successful", "success");
-            request.setAttribute("message", message);
+            if(reviewDAO.findByCustomer(customerId).size() > 0){
+                Message message = new Message("Delete failure", "Could not delete customer (ID: " + customerId
+                        + " ) because it currently contains some reviews", "error");
+                request.setAttribute("message", message);
+            }else if (customer.getBookOrdersByCustomerId().size() > 0){
+                Message message = new Message("Delete failure", "Could not delete customer (ID: " + customerId
+                        + " ) because it currently contains some orders", "error");
+                request.setAttribute("message", message);
+            } else {
+                customerDAO.delete(customerId);
+                Message message = new Message("Delete successful", "Delete customer successful", "success");
+                request.setAttribute("message", message);
+            }
+
         }
     }
